Toolbar now uses context api and tools can now be defined within their own route index instead of all in one place
