+++ interface/app/$libraryId/TopBar/TopBarMobile.tsx
@@ -1,23 +1,16 @@
 import { DotsThreeCircle } from 'phosphor-react';
-import { HTMLAttributes } from 'react';
-import { useLocation } from 'react-router-dom';
+import { HTMLAttributes, useContext } from 'react';
 import { Popover } from '@sd/ui';
+import { TOP_BAR_ICON_STYLE } from './ToolBarProvider';
+import { ToolBarContext } from './ToolBarProvider';
 import TopBarButton from './TopBarButton';
-import {
-	RoutePaths,
-	TOP_BAR_ICON_STYLE,
-	ToolOption,
-	useToolBarRouteOptions
-} from './useToolBarOptions';
 
 interface Props extends HTMLAttributes<HTMLDivElement> {}
 
 export default ({ className = '' }: Props) => {
-	const { pathname } = useLocation();
-	const getPageName = pathname.split('/')[2] as RoutePaths;
-	const { toolBarRouteOptions } = useToolBarRouteOptions();
-	const toolsNotSmFlex = toolBarRouteOptions[getPageName].options.map((group) =>
-		(group as ToolOption[]).filter((tool) => tool.showAtResolution !== 'sm:flex')
+	const { toolBar } = useContext(ToolBarContext);
+	const toolsNotSmFlex = toolBar.options.map((group) =>
+		group.filter((tool) => tool.showAtResolution !== 'sm:flex')
 	);
 
 	return (
@@ -31,12 +24,12 @@ export default ({ className = '' }: Props) => {
 			>
 				<div className="flex flex-col overflow-hidden p-2">
 					{toolsNotSmFlex.map((group, groupIndex) => {
-						return (group as ToolOption[]).map(
+						return group.map(
 							(
 								{ icon, onClick, popOverComponent, toolTipLabel, topBarActive },
 								index
 							) => {
-								const groupCount = toolBarRouteOptions[getPageName].options.length;
+								const groupCount = toolBar.options.length;
 								return (
 									<div key={toolTipLabel}>
 										{popOverComponent ? (
