+++ interface/app/$libraryId/location/$id.tsx
@@ -1,8 +1,24 @@
+import clsx from 'clsx';
+import {
+	ArrowClockwise,
+	Columns,
+	Key,
+	MonitorPlay,
+	Rows,
+	SidebarSimple,
+	SlidersHorizontal,
+	SquaresFour,
+	Tag
+} from 'phosphor-react';
 import { useEffect } from 'react';
 import { useParams, useSearchParams } from 'react-router-dom';
 import { useLibraryMutation, useLibraryQuery } from '@sd/client';
-import { getExplorerStore } from '~/hooks/useExplorerStore';
+import { getExplorerStore, useExplorerStore } from '~/hooks/useExplorerStore';
 import Explorer from '../Explorer';
+import OptionsPanel from '../Explorer/OptionsPanel';
+import { KeyManager } from '../KeyManager';
+import { TOP_BAR_ICON_STYLE, ToolOption } from '../TopBar/ToolBarProvider';
+import { useToolBar } from '../TopBar/useToolBar';
 
 export function useExplorerParams() {
 	const { id } = useParams<{ id?: string }>();
@@ -16,13 +32,95 @@ export function useExplorerParams() {
 }
 
 export const Component = () => {
+	const store = useExplorerStore();
+	const toolBarOptions: { options: ToolOption[][] } = {
+		options: [
+			[
+				{
+					toolTipLabel: 'Grid view',
+					icon: <SquaresFour className={TOP_BAR_ICON_STYLE} />,
+					topBarActive: store.layoutMode === 'grid',
+					onClick: () => (getExplorerStore().layoutMode = 'grid'),
+					showAtResolution: 'sm:flex'
+				},
+				{
+					toolTipLabel: 'List view',
+					icon: <Rows className={TOP_BAR_ICON_STYLE} />,
+					topBarActive: store.layoutMode === 'rows',
+					onClick: () => (getExplorerStore().layoutMode = 'rows'),
+					showAtResolution: 'sm:flex'
+				},
+				{
+					toolTipLabel: 'Columns view',
+					icon: <Columns className={TOP_BAR_ICON_STYLE} />,
+					topBarActive: store.layoutMode === 'columns',
+					onClick: () => (getExplorerStore().layoutMode = 'columns'),
+					showAtResolution: 'sm:flex'
+				},
+				{
+					toolTipLabel: 'Media view',
+					icon: <MonitorPlay className={TOP_BAR_ICON_STYLE} />,
+					topBarActive: store.layoutMode === 'media',
+					showAtResolution: 'sm:flex'
+				}
+			],
+			[
+				{
+					toolTipLabel: 'Key Manager',
+					icon: <Key className={TOP_BAR_ICON_STYLE} />,
+					popOverComponent: <KeyManager />,
+					individual: true,
+					showAtResolution: 'xl:flex'
+				},
+				{
+					toolTipLabel: 'Tag Assign Mode',
+					icon: (
+						<Tag
+							weight={store.tagAssignMode ? 'fill' : 'regular'}
+							className={TOP_BAR_ICON_STYLE}
+						/>
+					),
+					onClick: () => (getExplorerStore().tagAssignMode = !store.tagAssignMode),
+					topBarActive: store.tagAssignMode,
+					individual: true,
+					showAtResolution: 'xl:flex'
+				},
+				{
+					toolTipLabel: 'Regenerate thumbs (temp)',
+					icon: <ArrowClockwise className={TOP_BAR_ICON_STYLE} />,
+					individual: true,
+					showAtResolution: 'xl:flex'
+				}
+			],
+			[
+				{
+					toolTipLabel: 'Explorer display',
+					icon: <SlidersHorizontal className={TOP_BAR_ICON_STYLE} />,
+					popOverComponent: <OptionsPanel />,
+					individual: true,
+					showAtResolution: 'xl:flex'
+				},
+				{
+					toolTipLabel: 'Show Inspector',
+					onClick: () => (getExplorerStore().showInspector = !store.showInspector),
+					icon: (
+						<SidebarSimple
+							weight={store.showInspector ? 'fill' : 'regular'}
+							className={clsx(TOP_BAR_ICON_STYLE, 'scale-x-[-1]')}
+						/>
+					),
+					individual: true,
+					showAtResolution: 'xl:flex'
+				}
+			]
+		]
+	};
+	useToolBar(toolBarOptions);
 	const { location_id, path, limit } = useExplorerParams();
-
 	// we destructure this since `mutate` is a stable reference but the object it's in is not
 	const { mutate: mutateQuickRescan, ...quickRescan } =
 		useLibraryMutation('locations.quickRescan');
 	const explorerState = getExplorerStore();
-
 	useEffect(() => {
 		explorerState.locationId = location_id;
 		if (location_id !== null) mutateQuickRescan({ location_id, sub_path: path });
