+++ interface/app/$libraryId/overview.tsx
@@ -17,6 +17,8 @@ import { Statistics, useLibraryContext, useLibraryQuery } from '@sd/client';
 import { Card, ScreenHeading } from '@sd/ui';
 import useCounter from '~/hooks/useCounter';
 import { usePlatform } from '~/util/Platform';
+import { ToolOption } from './TopBar/ToolBarProvider';
+import { useToolBar } from './TopBar/useToolBar';
 
 interface StatItemProps {
 	title: string;
@@ -57,6 +59,10 @@ const StatItem = (props: StatItemProps) => {
 		duration: overviewMounted ? 0 : 1,
 		saveState: false
 	});
+	const toolBarOptions: { options: ToolOption[][] } = {
+		options: [[]]
+	};
+	useToolBar(toolBarOptions);
 
 	return (
 		<div
